Move Profile components to components folder
